+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod bar;
 pub mod blocks;
+pub mod common;
 pub mod config;
 pub mod os;
 pub mod sensors;